Move custom wood type registration back to ModContent
but it's slightly different from before as we don't depend on Neo's
RegisterEvent or RegisterHelper.
